linux-gen: ring: move ring headers into a separate directory
Move ring headers into a separate directory, so they don't  clutter the
main header directory.
