+++ platform/linux-generic/include/odp_queue_basic_internal.h
@@ -22,9 +22,9 @@ extern "C" {
 #include <odp/api/ticketlock.h>
 #include <odp_config_internal.h>
 #include <odp_macros_internal.h>
-#include <odp_ring_mpmc_ptr_internal.h>
-#include <odp_ring_st_ptr_internal.h>
-#include <odp_ring_spsc_ptr_internal.h>
+#include <ring/odp_ring_mpmc_ptr_internal.h>
+#include <ring/odp_ring_st_ptr_internal.h>
+#include <ring/odp_ring_spsc_ptr_internal.h>
 #include <odp_queue_lf.h>
 
 #define QUEUE_STATUS_FREE         0
