--- platform/linux-generic/include/odp_ring_st_ptr_internal.h
@@ -1,15 +1,0 @@
-/* SPDX-License-Identifier: BSD-3-Clause
- * Copyright (c) 2025 Nokia
- */
-
-#ifndef ODP_RING_ST_PTR_INTERNAL_H_
-#define ODP_RING_ST_PTR_INTERNAL_H_
-
-#include <odp_ring_common.h>
-
-#undef _ODP_RING_TYPE
-#define _ODP_RING_TYPE _ODP_RING_TYPE_PTR
-
-#include <odp_ring_st_internal.h>
-
-#endif
