+++ platform/linux-generic/include/ring/odp_ring_common.h
