+++ platform/linux-generic/include/ring/odp_ring_mpmc_rst_u64_internal.h
@@ -5,11 +5,11 @@
 #ifndef ODP_RING_MPMC_RST_U64_INTERNAL_H_
 #define ODP_RING_MPMC_RST_U64_INTERNAL_H_
 
-#include <odp_ring_common.h>
+#include <ring/odp_ring_common.h>
 
 #undef _ODP_RING_TYPE
 #define _ODP_RING_TYPE _ODP_RING_TYPE_U64
 
-#include <odp_ring_mpmc_rst_internal.h>
+#include <ring/odp_ring_mpmc_rst_internal.h>
 
 #endif
