+++ platform/linux-generic/include/ring/odp_ring_mpmc_u32_internal.h
@@ -9,12 +9,12 @@
 extern "C" {
 #endif
 
-#include <odp_ring_common.h>
+#include <ring/odp_ring_common.h>
 
 #undef _ODP_RING_TYPE
 #define _ODP_RING_TYPE _ODP_RING_TYPE_U32
 
-#include <odp_ring_mpmc_internal.h>
+#include <ring/odp_ring_mpmc_internal.h>
 
 #ifdef __cplusplus
 }
