+++ platform/linux-generic/include/ring/odp_ring_spsc_internal.h
