+++ platform/linux-generic/include/ring/odp_ring_spsc_ptr_internal.h
@@ -1,0 +1,15 @@
+/* SPDX-License-Identifier: BSD-3-Clause
+ * Copyright (c) 2025 Nokia
+ */
+
+#ifndef ODP_RING_SPSC_PTR_INTERNAL_H_
+#define ODP_RING_SPSC_PTR_INTERNAL_H_
+
+#include <ring/odp_ring_common.h>
+
+#undef _ODP_RING_TYPE
+#define _ODP_RING_TYPE _ODP_RING_TYPE_PTR
+
+#include <ring/odp_ring_spsc_internal.h>
+
+#endif
