+++ platform/linux-generic/include/ring/odp_ring_st_ptr_internal.h
@@ -2,14 +2,14 @@
  * Copyright (c) 2025 Nokia
  */
 
-#ifndef ODP_RING_SPSC_PTR_INTERNAL_H_
-#define ODP_RING_SPSC_PTR_INTERNAL_H_
+#ifndef ODP_RING_ST_PTR_INTERNAL_H_
+#define ODP_RING_ST_PTR_INTERNAL_H_
 
-#include <odp_ring_common.h>
+#include <ring/odp_ring_common.h>
 
 #undef _ODP_RING_TYPE
 #define _ODP_RING_TYPE _ODP_RING_TYPE_PTR
 
-#include <odp_ring_spsc_internal.h>
+#include <ring/odp_ring_st_internal.h>
 
 #endif
