+++ platform/linux-generic/pktio/ipc.c
@@ -12,7 +12,7 @@
 #include <odp_pool_internal.h>
 #include <odp_macros_internal.h>
 #include <odp_shm_internal.h>
-#include <odp_ring_mpmc_rst_ptr_internal.h>
+#include <ring/odp_ring_mpmc_rst_ptr_internal.h>
 #include <odp_global_data.h>
 
 #include <fcntl.h>
